Task-2_Bank_Statement_Parser/requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pdf2image>=1.16.3
pdfminer.six>=20221105
google-genai>=0.2.0
pandas>=2.0.0
